requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
gunicorn==20.1.0
httpie==2.4.0
idna==2.10
iniconfig==1.1.1
lml==0.1.0
Markdown==3.3.4
MarkupPy==1.14
numpy==1.20.2
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
pluggy==0.13.1
py==1.10.0
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
Pygments==2.8.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
